+++ tests/test_access.py
@@ -1,0 +1,242 @@
+#  Copyright 2018 Ocean Protocol Foundation
+#  SPDX-License-Identifier: Apache-2.0
+
+import json
+import mimetypes
+from unittest.mock import Mock, MagicMock
+import uuid
+
+from eth_utils import add_0x_prefix
+from ocean_keeper.utils import add_ethereum_prefix_and_hash_msg
+from ocean_utils.agreements.service_agreement import ServiceAgreement
+from ocean_utils.agreements.service_types import ServiceTypes
+from ocean_utils.http_requests.requests_session import get_requests_session
+from werkzeug.utils import get_content_type
+
+from ocean_utils.did import DID, did_to_id
+
+from tests.conftest import get_sample_ddo
+
+from brizo.constants import BaseURLs
+from brizo.util import (check_auth_token, do_secret_store_decrypt, do_secret_store_encrypt,
+                        generate_token, get_config, get_provider_account, is_token_valid,
+                        keeper_instance,
+                        verify_signature,
+                        web3,
+                        build_download_response, get_download_url, get_latest_keeper_version)
+
+from tests.conftest import get_consumer_account, get_publisher_account
+
+from tests.test_helpers import (get_access_service_descriptor, get_compute_service_descriptor, get_dataset_ddo_with_access_service, \
+                                get_registered_ddo, get_dataset_ddo_with_compute_service, lock_reward, place_order, grant_access)
+
+PURCHASE_ENDPOINT = BaseURLs.BASE_BRIZO_URL + '/services/access/initialize'
+SERVICE_ENDPOINT = BaseURLs.BASE_BRIZO_URL + '/services/consume'
+
+
+def dummy_callback(*_):
+    pass
+
+def test_consume(client):
+    endpoint = BaseURLs.ASSETS_URL + '/consume'
+
+    pub_acc = get_publisher_account()
+    cons_acc = get_consumer_account()
+
+    keeper = keeper_instance()
+    ddo = get_dataset_ddo_with_access_service(pub_acc, providers=[pub_acc.address])
+
+    # initialize an agreement
+    agreement_id = place_order(pub_acc, ddo, cons_acc, ServiceTypes.ASSET_ACCESS)
+    payload = dict({
+        'serviceAgreementId': agreement_id,
+        'consumerAddress': cons_acc.address
+    })
+
+    agr_id_hash = add_ethereum_prefix_and_hash_msg(agreement_id)
+    signature = keeper.sign_hash(agr_id_hash, cons_acc)
+    index = 0
+
+    event = keeper.agreement_manager.subscribe_agreement_created(
+        agreement_id, 15, None, (), wait=True, from_block=0
+    )
+    assert event, "Agreement event is not found, check the keeper node's logs"
+
+    consumer_balance = keeper.token.get_token_balance(cons_acc.address)
+    if consumer_balance < 50:
+        keeper.request_tokens(50-consumer_balance, cons_acc)
+
+    sa = ServiceAgreement.from_ddo(ServiceTypes.ASSET_ACCESS, ddo)
+    lock_reward(agreement_id, sa, cons_acc)
+    event = keeper.lock_reward_condition.subscribe_condition_fulfilled(
+        agreement_id, 15, None, (), wait=True, from_block=0
+    )
+    assert event, "Lock reward condition fulfilled event is not found, check the keeper node's logs"
+
+    grant_access(agreement_id, ddo, cons_acc, pub_acc)
+    event = keeper.access_secret_store_condition.subscribe_condition_fulfilled(
+        agreement_id, 15, None, (), wait=True, from_block=0
+    )
+    assert event or keeper.access_secret_store_condition.check_permissions(
+        ddo.asset_id, cons_acc.address
+    ), f'Failed to get access permission: agreement_id={agreement_id}, ' \
+       f'did={ddo.did}, consumer={cons_acc.address}'
+
+    # Consume using decrypted url
+    files_list = json.loads(
+        do_secret_store_decrypt(did_to_id(ddo.did), ddo.encrypted_files, pub_acc, get_config()))
+    payload['url'] = files_list[index]['url']
+    request_url = endpoint + '?' + '&'.join([f'{k}={v}' for k, v in payload.items()])
+
+    response = client.get(
+        request_url
+    )
+    assert response.status == '200 OK'
+
+    # Consume using url index and signature (let brizo do the decryption)
+    payload.pop('url')
+    payload['signature'] = signature
+    payload['index'] = index
+    request_url = endpoint + '?' + '&'.join([f'{k}={v}' for k, v in payload.items()])
+    response = client.get(
+        request_url
+    )
+    assert response.status == '200 OK'
+
+
+def test_empty_payload(client):
+    consume = client.get(
+        BaseURLs.ASSETS_URL + '/consume',
+        data=None,
+        content_type='application/json'
+    )
+    assert consume.status_code == 400
+
+    publish = client.post(
+        BaseURLs.ASSETS_URL + '/publish',
+        data=None,
+        content_type='application/json'
+    )
+    assert publish.status_code == 400
+
+
+def test_publish(client):
+    endpoint = BaseURLs.ASSETS_URL + '/publish'
+    did = DID.did({"0": str(uuid.uuid4())})
+    asset_id = did_to_id(did)
+    account = get_provider_account()
+    test_urls = [
+        'url 00',
+        'url 11',
+        'url 22'
+    ]
+    keeper = keeper_instance()
+    urls_json = json.dumps(test_urls)
+    asset_id_hash = add_ethereum_prefix_and_hash_msg(asset_id)
+    signature = keeper.sign_hash(asset_id_hash, account)
+    address = web3().eth.account.recoverHash(asset_id_hash, signature=signature)
+    assert address.lower() == account.address.lower()
+    address = keeper.personal_ec_recover(asset_id, signature)
+    assert address.lower() == account.address.lower()
+
+    payload = {
+        'documentId': asset_id,
+        'signature': signature,
+        'document': urls_json,
+        'publisherAddress': account.address
+    }
+    post_response = client.post(
+        endpoint,
+        data=json.dumps(payload),
+        content_type='application/json'
+    )
+    encrypted_url = post_response.data.decode('utf-8')
+    assert encrypted_url.startswith('0x')
+
+    # publish using auth token
+    signature = generate_token(account)
+    payload['signature'] = signature
+    did = DID.did({"0": str(uuid.uuid4())})
+    asset_id = did_to_id(did)
+    payload['documentId'] = add_0x_prefix(asset_id)
+    post_response = client.post(
+        endpoint,
+        data=json.dumps(payload),
+        content_type='application/json'
+    )
+    encrypted_url = post_response.data.decode('utf-8')
+    assert encrypted_url.startswith('0x')
+
+
+def test_auth_token():
+    token = "0x1d2741dee30e64989ef0203957c01b14f250f5d2f6ccb0" \
+            "c88c9518816e4fcec16f84e545094eb3f377b7e214ded226" \
+            "76fbde8ca2e41b4eb1b3565047ecd9acf300-1568372035"
+    pub_address = "0xe2DD09d719Da89e5a3D0F2549c7E24566e947260"
+    doc_id = "663516d306904651bbcf9fe45a00477c215c7303d8a24c5bad6005dd2f95e68e"
+    assert is_token_valid(token), f'cannot recognize auth-token {token}'
+    address = check_auth_token(token)
+    assert address and address.lower() == pub_address.lower(), f'address mismatch, got {address}, ' \
+                                                               f'' \
+                                                               f'' \
+                                                               f'expected {pub_address}'
+    good = verify_signature(keeper_instance(), pub_address, token, doc_id)
+    assert good, f'invalid signature/auth-token {token}, {pub_address}, {doc_id}'
+
+
+def test_exec_endpoint():
+    pass
+
+
+def test_download_ipfs_file(client):
+    cid = 'QmQfpdcMWnLTXKKW9GPV7NgtEugghgD6HgzSF6gSrp2mL9'
+    url = f'ipfs://{cid}'
+    download_url = get_download_url(url, None)
+    requests_session = get_requests_session()
+
+    request = Mock()
+    request.range = None
+
+    print(f'got ipfs download url: {download_url}')
+    assert download_url and download_url.endswith(f'ipfs/{cid}')
+    response = build_download_response(request, requests_session, download_url, download_url, None)
+    assert response.data, f'got no data {response.data}'
+
+
+def test_build_download_response():
+    request = Mock()
+    request.range = None
+
+    class Dummy:
+        pass
+
+    response = Dummy()
+    response.content = b'asdsadf'
+    response.status_code = 200
+
+    requests_session = Dummy()
+    requests_session.get = MagicMock(return_value=response)
+
+    filename = '<<filename>>.xml'
+    content_type = mimetypes.guess_type(filename)[0]
+    url = f'https://source-lllllll.cccc/{filename}'
+    response = build_download_response(request, requests_session, url, url, None)
+    assert response.headers["content-type"] == content_type
+    assert response.headers.get_all('Content-Disposition')[0] == f'attachment;filename={filename}'
+
+    filename = '<<filename>>'
+    url = f'https://source-lllllll.cccc/{filename}'
+    response = build_download_response(request, requests_session, url, url, None)
+    assert response.headers["content-type"] == get_content_type(response.default_mimetype, response.charset)
+    assert response.headers.get_all('Content-Disposition')[0] == f'attachment;filename={filename}'
+
+    filename = '<<filename>>'
+    url = f'https://source-lllllll.cccc/{filename}'
+    response = build_download_response(request, requests_session, url, url, content_type)
+    assert response.headers["content-type"] == content_type
+    assert response.headers.get_all('Content-Disposition')[0] == f'attachment;filename={filename+mimetypes.guess_extension(content_type)}'
+
+
+def test_latest_keeper_version():
+    version = get_latest_keeper_version()
+    assert version.startswith('v') and len(version.split('.')) == 3, ''
